refactor: switched file name convention from PascalCase to kebab-case
