+++ apps/web/app/(public)/layout.tsx
@@ -1,6 +1,6 @@
 import { Fragment, type ReactNode } from "react";
 
-import { Footer } from "@/components/Footer";
+import { Footer } from "@/components/footer";
 
 export default function AppLayout({
 	children,
