+++ apps/web/app/account/settings/customization/page.tsx
@@ -1,4 +1,4 @@
-import { ToDo } from "@/components/ToDo";
+import { ToDo } from "@/components/to-do";
 
 export default function CustomizationSettings() {
 	return <ToDo />;
