+++ apps/web/app/account/settings/language/page.tsx
@@ -1,4 +1,4 @@
-import { ToDo } from "@/components/ToDo";
+import { ToDo } from "@/components/to-do";
 
 export default function LanguageSettings() {
 	return <ToDo />;
