+++ apps/web/app/account/settings/layout.tsx
@@ -1,4 +1,4 @@
-import { SettingsSidebar } from "@/components/settings/Sidebar";
+import { SettingsSidebar } from "@/components/settings/sidebar";
 
 const settingsLinks = [
 	{
@@ -22,13 +22,10 @@ export default function AccountSettingsLayout({
 }) {
 	return (
 		<main className="container mx-auto py-8 px-4 max-w-6xl min-h-screen relative">
-			<h1 className="text-3xl font-bold text-[#333333] mb-8">Configurações</h1>
+			<h1 className="text-3xl font-bold text-foreground mb-8">Configurações</h1>
 
 			<div className="flex flex-col md:flex-row gap-8 h-full">
-				{/* Sidebar */}
 				<SettingsSidebar links={settingsLinks} />
-
-				{/* Main content */}
 				<div className="flex-1">{children}</div>
 			</div>
 		</main>
