+++ apps/web/app/dashboard/participants/@modal/(..)members/[memberId]/page.tsx
@@ -1,6 +1,6 @@
 // import { ParticipantCard, ParticipantCardModal } from "@/components/participants/ParticipantCard";
 
-import { ToDo } from "@/components/ToDo";
+import { ToDo } from "@/components/to-do";
 
 export default async function ParticipantModal(props: {
 	params: Promise<{ participantId: string }>;
