+++ apps/web/hooks/use-mobile.tsx
@@ -1,0 +1,26 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+export function useIsMobile() {
+	const [isMobile, setIsMobile] = useState(false);
+
+	useEffect(() => {
+		const checkIfMobile = () => {
+			setIsMobile(window.innerWidth < 768);
+		};
+
+		// Initial check
+		checkIfMobile();
+
+		// Add event listener
+		window.addEventListener("resize", checkIfMobile);
+
+		// Clean up
+		return () => {
+			window.removeEventListener("resize", checkIfMobile);
+		};
+	}, []);
+
+	return isMobile;
+}
